Fix for multisell
git-svn-id: http://svn.l2jserver.com/trunk@3211 d34dfa0f-f91c-0410-86dd-cf06f19dd0ad

+++ L2_GameServer/java/net/sf/l2j/gameserver/model/L2Multisell.java
@@ -710,7 +710,7 @@ public class L2Multisell
 				Node attribute;
 				
 				int id = Integer.parseInt(n.getAttributes().getNamedItem("id").getNodeValue());
-				long count = Integer.parseInt(n.getAttributes().getNamedItem("count").getNodeValue());
+				long count = Long.parseLong(n.getAttributes().getNamedItem("count").getNodeValue());
 				boolean isTaxIngredient = false, mantainIngredient = false;
 				
 				attribute = n.getAttributes().getNamedItem("isTaxIngredient");
@@ -729,7 +729,7 @@ public class L2Multisell
 			else if ("production".equalsIgnoreCase(n.getNodeName()))
 			{
 				int id = Integer.parseInt(n.getAttributes().getNamedItem("id").getNodeValue());
-				long count = Integer.parseInt(n.getAttributes().getNamedItem("count").getNodeValue());
+				long count = Long.parseLong(n.getAttributes().getNamedItem("count").getNodeValue());
 				
 				MultiSellIngredient e = new MultiSellIngredient(id, count, false, false);
 				entry.addProduct(e);
